Separate general features for numerical integrator
